[feat] 정답체크 List형태로 Body받게 변경 + 리턴 타입 int 로 변경 (정답시 1/ 오답시 0)

+++ eTest/src/main/java/com/tmax/eTest/Contents/repository/ProblemRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 
 import com.tmax.eTest.Contents.model.Problem;
 
-public interface ProblemRepository extends JpaRepository<Problem, Long>{
+public interface ProblemRepository extends JpaRepository<Problem, Integer>{
 
 }
